backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.1
pydantic==1.10.11
motor==3.7.1
python-dotenv==1.0.1
openai==1.40.6
reportlab==4.0.8
httpx>=0.27.0
